+++ ikea-zigbee-drivers/Ikea_E2013.groovy
@@ -66,6 +66,16 @@ metadata {
             required: true
         )
         
+        // Inputs for devices.E2013
+        input(
+            name: 'swapOpenClosed',
+            type: 'bool',
+            title: 'Swap open / closed',
+            description: '<small>Swap open / closed value for the "contact" attribute.</small>',
+            defaultValue: false,
+            required: true
+        )
+        
         // Inputs for capability.ZigbeeBindings
         input(
             name: 'controlDevice',
@@ -111,6 +121,13 @@ List<String> updated(boolean auto = false) {
     }
     if (logLevel == '1') runIn 1800, 'logsOff'
     log_info "🛠️ logLevel = ${['1':'Debug', '2':'Info', '3':'Warning', '4':'Error'].get(logLevel)}"
+    
+    // Preferences for devices.E2013
+    if (swapOpenClosed == null) {
+        swapOpenClosed = false
+        device.updateSetting('swapOpenClosed', [value:swapOpenClosed, type:'bool'])
+    }
+    log_info "🛠️ swapOpenClosed = ${swapOpenClosed}"
     
     // Preferences for capability.HealthCheck
     schedule HEALTH_CHECK.schedule, 'healthCheck'
@@ -339,7 +356,7 @@ void parse(String description) {
         // Report/Read Attributes Reponse: ZoneStatus
         case { contains it, [clusterInt:0x0500, commandInt:0x0A, attrInt:0x0002] }:
         case { contains it, [clusterInt:0x0500, commandInt:0x01, attrInt:0x0002] }:
-            String contact = msg.value[-1] == '1' ? 'open' : 'closed'
+            String contact = msg.value[-1] == '1' ^ (swapOpenClosed == true) ? 'open' : 'closed'
             utils_sendEvent name:'contact', value:contact, descriptionText:"Is ${contact}", type:type
             utils_processedZclMessage "${msg.commandInt == 0x0A ? 'Report' : 'Read'} Attributes Response", "ZoneStatus=${msg.value}"
             return