+++ ikea-zigbee-drivers/src/devices/Aqara_DCM-K01/DCM-K01.groovy
@@ -82,7 +82,7 @@ input(
     options: [
         '0':'Disabled - control lights and other devices',
         '1':'Enabled - control bi-directional motors',
-    ],,
+    ],
     defaultValue: '0',
     required: true
 )