+++ ikea-zigbee-drivers/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add Zigbee Groups membership control for mains-powered devices (E1603, E1836, ICPSHC24)
 - Add Zigbee Bindings control for devices that support this feature (E1743, E1810, E1812, E2002, E2013, E2123, E2201, RDM001, RWL022)
+- E2013: App option to swap "open" / "closed" value for the "contact" attribute
 - Add driver for Aqara Dual Relay Module T2 (DCM-K01)
 - Add driver for Philips Hue Wall Switch Module (RDM001)
 - Add driver for Philips Hue Dimmer Switch (RWL022)